compatibility for Stud.IP v3.1 - v4.0 and general facelift

+++ app/views/index/_meetings.php
@@ -100,13 +100,22 @@ if ($showUser) {
                     </a>
                     <input type="text" name="name"><br>
                     <input type="text" name="recording_url" placeholder="<?=_('URL zur Aufzeichnung')?>">
-                    <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/grey/accept.png" class="accept-button" title="<?=_('Änderungen speichern')?>">
-                    <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/grey/decline.png" class="decline-button" title="<?=_('Änderungen verwerfen')?>">
+                    <? if (StudipVersion::newerThan('3.3')) : ?>
+                        <?= Icon::create('accept', 'clickable', array('class' => 'accept-button', 'title' => _('Änderungen speichern'))) ?>
+                        <?= Icon::create('decline', 'clickable', array('class' => 'decline-button', 'title' => _('Änderungen verwerfen'))) ?>
+                    <? else: ?>
+                        <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/accept.png" class="accept-button" title="<?=_('Änderungen speichern')?>">
+                        <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/decline.png" class="decline-button" title="<?=_('Änderungen verwerfen')?>">
+                    <? endif ?>
                     <img src="<?=$GLOBALS['ASSETS_URL']?>/images/ajax_indicator_small.gif" class="loading-indicator">
                 </td>
                 <td class="recording-url">
                     <a href="<?=$meetingCourse->meeting->recording_url?>" target="_blank" class="meeting-recording-url"<?=!$meetingCourse->meeting->recording_url ? ' style="display:none;"' : ''?>>
-                        <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/blue/video.png" title="<?=_('zur Aufzeichnung')?>">
+                        <? if (StudipVersion::newerThan('3.3')) : ?>
+                            <?= Icon::create('video', 'clickable', array('title' => _('zur Aufzeichnung'))) ?>
+                        <? else: ?>
+                            <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/video.png" title="<?=_('zur Aufzeichnung')?>">
+                        <? endif ?>
                     </a>
                 </td>
                 <?php if ($showCourse): ?>
@@ -137,22 +146,54 @@ if ($showUser) {
                     </td>
                     <td class="active"><input type="checkbox"<?=$meetingCourse->active ? ' checked="checked"' : ''?> data-meeting-enable-url="<?=PluginEngine::getLink($plugin, array('destination' => $destination), 'index/enable/'.$meetingCourse->meeting->id.'/'.$meetingCourse->course->id)?>" title="<?=$meetingCourse->active ? _('Meeting für Teilnehmende unsichtbar schalten') : _('Meeting für Teilnehmende sichtbar schalten')?>"></td>
                     <td>
-                        <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/blue/info-circle.png" title="<?=_('Informationen anzeigen')?>" class="info">
-                        <a href="#" title="<?=_('Meeting bearbeiten')?>" class="edit-meeting" data-meeting-edit-url="<?=PluginEngine::getLink($plugin, array(), 'index/edit/'.$meetingCourse->meeting->id)?>"><img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/blue/edit.png"></a>
-                        <?php if ($meetingCourse->meeting->join_as_moderator): ?>
-                            <a href="<?=$moderatorPermissionsUrl?>" title="<?=_('Teilnehmende haben VeranstalterInnen-Rechte')?>"><img src="<?=$plugin->getAssetsUrl()?>/images/moderator-enabled.png"></a>
-                        <?php else: ?>
-                            <a href="<?=$moderatorPermissionsUrl?>" title="<?=_('Teilnehmende haben eingeschränkte Rechte')?>"><img src="<?=$plugin->getAssetsUrl()?>/images/moderator-disabled.png"></a>
-                        <?php endif; ?>
-                        <a href="<?=$deleteUrl?>" title="<?=count($meetingCourse->meeting->courses) > 1 ? _('Zuordnung löschen') : _('Meeting löschen')?>">
-                            <?php if (count($meetingCourse->meeting->courses) > 1): ?>
-                                <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/blue/remove.png">
-                            <?php else: ?>
-                                <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/20/blue/trash.png">
-                            <?php endif ?>
-                        </a>
+                        <? if (StudipVersion::newerThan('3.3')) : ?>
+                            <?= Icon::create('info-circle', 'clickable', array('class' => 'info')) ?>
+                            <a href="#" title="<?=_('Meeting bearbeiten')?>" class="edit-meeting" data-meeting-edit-url="<?=PluginEngine::getLink($plugin, array(), 'index/edit/'.$meetingCourse->meeting->id)?>">
+                                <?= Icon::create('edit') ?>
+                            </a>
+                            <? if ($meetingCourse->meeting->join_as_moderator): ?>
+                                <a href="<?= $moderatorPermissionsUrl ?>" title="<?=_('Teilnehmende haben VeranstalterInnen-Rechte')?>">
+                                    <?= Icon::create('admin') ?>
+                                </a>
+                            <? else: ?>
+                                <a href="<?= $moderatorPermissionsUrl ?>" title="<?=_('Teilnehmende haben eingeschränkte Rechte')?>">
+                                    <?= Icon::create('admin+decline') ?>
+                                </a>
+                            <? endif; ?>
+
+                            <a href="<?= $deleteUrl ?>" title="<?= count($meetingCourse->meeting->courses) > 1 ? _('Zuordnung löschen') : _('Meeting löschen') ?>">
+                                <? if (count($meetingCourse->meeting->courses) > 1): ?>
+                                    <?= Icon::create('remove') ?>
+                                <? else: ?>
+                                    <?= Icon::create('trash') ?>
+                                <? endif ?>
+                            </a>
+                        <? else : ?>
+                            <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/info-circle.png" title="<?=_('Informationen anzeigen')?>" class="info">
+                            <a href="#" title="<?=_('Meeting bearbeiten')?>" class="edit-meeting" data-meeting-edit-url="<?=PluginEngine::getLink($plugin, array(), 'index/edit/'.$meetingCourse->meeting->id)?>">
+                                <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/edit.png">
+                            </a>
+
+                            <? if ($meetingCourse->meeting->join_as_moderator): ?>
+                                <a href="<?= $moderatorPermissionsUrl ?>" title="<?=_('Teilnehmende haben VeranstalterInnen-Rechte')?>">
+                                    <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/admin.png">
+                                </a>
+                            <? else: ?>
+                                <a href="<?= $moderatorPermissionsUrl ?>" title="<?=_('Teilnehmende haben eingeschränkte Rechte')?>">
+                                    <img src="<?=$plugin->getAssetsUrl()?>/images/admin-decline.png">
+                                </a>
+                            <? endif; ?>
+
+                            <a href="<?= $deleteUrl ?>" title="<?= count($meetingCourse->meeting->courses) > 1 ? _('Zuordnung löschen') : _('Meeting löschen') ?>">
+                                <? if (count($meetingCourse->meeting->courses) > 1): ?>
+                                    <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/remove.png">
+                                <? else: ?>
+                                    <img src="<?=$GLOBALS['ASSETS_URL']?>/images/icons/16/blue/trash.png">
+                                <? endif ?>
+                            </a>
+                        <? endif ?>
                     </td>
-                <?php endif; ?>
+                <? endif; ?>
                 </tr>
 
                 <?php if ($canModifyMeetings): ?>
@@ -186,9 +227,9 @@ if ($showUser) {
         <?php if ($canModifyMeetings): ?>
             <tfoot>
             <tr>
-                <td colspan="<?=$colspan?>">
-                    <input class="middle" type="checkbox" name="check_all" title="<?=_('Alle Meetings auswählen')?>">
-                    <?=Studip\Button::create(_('Löschen'), array('title' => _('Alle ausgewählten Meetings löschen')))?>
+                <td colspan="<?= $colspan ?>">
+                    <input class="middle" type="checkbox" name="check_all" title="<?= _('Alle Meetings auswählen') ?>">
+                    <?= Studip\Button::create(_('Löschen'), array('title' => _('Alle ausgewählten Meetings löschen'))) ?>
                 </td>
             </tr>
             </tfoot>
